+++ bpod_core/bpod/abc.py
@@ -335,7 +335,11 @@ class FlexIOChannel:
 class AbstractFlexIO(SuggestionMapping[FlexIOChannel]):
     """Abstract base for FlexIO subsystems."""
 
-    __slots__ = ('_on_channel_types_changed', '_state')
+    __slots__ = (
+        '_on_analog_sampling_rate_changed',
+        '_on_channel_types_changed',
+        '_state',
+    )
 
     _state: _FlexIOState
 
@@ -344,9 +348,11 @@ class AbstractFlexIO(SuggestionMapping[FlexIOChannel]):
         n: int,
         *,
         on_channel_types_changed: Callable[[], None] | None = None,
+        on_analog_sampling_rate_changed: Callable[[], None] | None = None,
     ) -> None:
         self._state = _FlexIOState.create_default(n_channels=n)
         self._on_channel_types_changed = on_channel_types_changed
+        self._on_analog_sampling_rate_changed = on_analog_sampling_rate_changed
         view = {}
         for i in range(n):
             channel = FlexIOChannel(self._get_state, self._apply_settings, i)